pyproject.toml
[project]
name = "memory-palace"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipython>=8.37.0",
    "pandas>=2.3.2",
    "pgvector>=0.4.1",
    "pika>=1.3.2",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
]
